backend/requirements.txt
flask==3.0.2
flask-cors==4.0.0
google-generativeai==0.3.2
flask-sqlalchemy==3.1.1
werkzeug==3.0.1
python-dotenv==1.0.1
pymongo==4.6.1
chromadb==0.4.15
# PostgreSQL migration dependencies
psycopg2-binary==2.9.9
pgvector==0.2.4
sqlalchemy==2.0.23
alembic==1.13.1
sentence-transformers>=2.7.0
redis==5.0.1
# RAG improvements
tiktoken==0.5.2
rank-bm25==0.2.2
huggingface-hub>=0.20.0
ollama==0.5.1